added start css for the footer
i will need to work on it more for accessibility for smaller screens. used grids to create the bottom, worked out very well with the help of a couple youtube video. for next time, i'm either going to start on the flash page or work on the @media for the footer